refactor: 스키마 분리
- 당근 도메인 및 생일 축하 도메인에 대한 스키마를 분리합니다
