+++ src/schema/index.ts
@@ -1,1 +1,3 @@
 export * from './schema';
+export * from './carrot.schema';
+export * from './birthday.schema';
